+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Vendoring/OpenTelemetry.Resources.Azure/AppServiceResourceDetector.cs
@@ -30,7 +30,11 @@ internal sealed class AppServiceResourceDetector : IResourceDetector
 
             if (websiteSiteName != null)
             {
-                attributeList.Add(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, websiteSiteName));
+                if (string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable(ResourceAttributeConstants.OpenTelemetryServiceNameEnvVar)))
+                {
+                    attributeList.Add(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, websiteSiteName));
+                }
+
                 attributeList.Add(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeCloudProvider, ResourceAttributeConstants.AzureCloudProviderValue));
                 attributeList.Add(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeCloudPlatform, ResourceAttributeConstants.AzureAppServicePlatformValue));